macOS: fix Apple Clang (libc++) build of opennn
Apple Clang 17's libc++ rejects code the Windows/Linux builds accept:

- std::jthread / <stop_token> (batch.h) are gated behind the experimental
  library: add -fexperimental-library (compile + link) in the APPLE block.
- std::from_chars / std::to_chars for floating point are deleted: add an
  __APPLE__ opennn::from_chars shim (integral -> std, float -> strtod/strtof)
  in string_utilities.h; include it from tabular_dataset.cpp.
- to_string(file_time_type ... .count()) is ambiguous: cast to long long
  in image_dataset.cpp.

+++ opennn/image_dataset.cpp
@@ -54,7 +54,7 @@ static string image_cache_signature(Index samples, Index height, Index width, In
 {
     string signature = to_string(samples) + "|"
         + to_string(height) + "x" + to_string(width) + "x" + to_string(channels) + "|"
-        + to_string(newest_write_time.time_since_epoch().count()) + "|";
+        + to_string(static_cast<long long>(newest_write_time.time_since_epoch().count())) + "|";
     for (const filesystem::path& folder : class_folders)
         signature += folder.filename().string() + ",";
     return signature;
